Added /Ac/../NumberOfPhases for all PV Inverter types
The maximum available phases determines the number of phases:
if L3 is available, number of phases = 3
if L2 is available, number of phases = 2
if L1 is available, number of phases = 1
if no phases are available, number of phases is invalid

In the future, when the processes that put the pvinverter info on the dbus
(dbus_qwacs and dbus_vebus_to_pvinverter) also have ../NumberOfPhases we need
to read that instead of doing this determination.

The functional difference is that now the NumberOfPhases shows how much phases
are available, and by looking at ../NumberOfPhases as available from qwacs for
example we show the number of phases as configured by the user.

The latter is what we want.