cope with double instances on startup
This came to light on Groningen Hub-1, where there is a vecan to
vebus interface installed (device instance 0). And also a multi
was connected directly to the vebus port (device instance 257).
Dbusmonitor in velib_python changes that last device instance to
0, which has a historic reason: in the VGR/VER time, all multis
were always at device instance 0 on the VRM Portal. And getting
them at different instances would have broke it.

But this causes double device instance 0 for vebus, and that
caused systemcalc to crash on startup.

Note that besides this change, also the code in Dbusmonitor that
changes all vebus device instances to 0 needs to be either moved
or changed to work only for vrmlogger, or even better: we finally
get rid of that exception. Since now the service map from system-
calc is wrong:
/ServiceMapping/com_victronenergy_pvinverter_11
/ServiceMapping/com_victronenergy_settings_0
/ServiceMapping/com_victronenergy_vebus_0  <-- WRONG, Is 257!

Which doesn't seem to cause to big issues yet, but for sure will
cause them.