Improved computation of consumption in incomplete systems.
Consumption is computed for each vebus connection (AcIn1,AcIn2,AcOut).
For each connection the absolute value of the total consumption is used.
Previously all consumption would be summed and thereafter the absolute
value was computed. This causes deviations is systems where a PV
inverter is present, but not connected to the CCGX. A negative consumption
would be calculated for associated connection, which would disturb the
consumption values computed for other connections.