Select vebus service as auto battery service in setup with solar charger
If a setup contains a vebus service and one or more solar chargers, but no
battery service  the vebus will have an accurate SoC due to the
ExtraBatteryCurrent parameter in the vebus service.

If this setup is detected we will select the vebus service as the auto
battery service. Note: there is a check on the presence of
/ExtraBatteryCurrent.