No longer copy the max charge current from BMS battery to solar charger.
Copying the max charge current from BMS batteries to the solarcharger leads to problems:
excess PV power is not fed back to the grid any more, and loads on AC-out are not fed with PV power.
PV power is used for charging the batteries only.

So we removed this feature, until we have a complete solution for solar charger support. Until then
we set a 'high' max charge current to avoid 'BMS connection lost' alarms from the solarcharger.