Always write solar charger current to vebus service.
The solar charger current used to be written to the vebus service to
improve the SoC computation. If there was an alternative SoC source
present in the system (BMV, CAN-bus BMS,...), that SoC would be copied
to the vebus service, and the solar charger current would not be copied.

It turned out the solar charger current should also be copied in that
situation to allow the Multi to compute the right duration of the
absorption stage.