Simplified copy that copies SoC to vebus
Solar charger current would be copied to the vebus service only
if no SoC was copied. Now the current is always copied, so we
can simplify the code somewhat.

We no longer need to cache whether or not a Hub-2 assistant is
present, because we only need it for SoC copying which runs on a
slow (10s) timer.