ESS: do not copy BMS max charge voltage to vebus
We don't do this, because the ESS assistant uses the
BOL max charge voltage to compute the voltage setpoint
for MPPTs. This may cause overvoltage errors in BMS
batteries because the ESS assistant sometimes adds an
offset (max 1V @12V battery, 4V @64V).

Without copying the max charge voltage the assistant will
use the absorption voltage instead of the BMS max voltage.
This allows the user to mitigate the problem by setting
an absorption voltage using VEConfig which is less than the
BMS max voltage.