Check vebus /Hub2 before copying SoC
Normally SoC is copied from the active battery monitor to the
vebus service. This is not allowed if a hub-2 assistant is
present in the vebus device. To check this the assistant IDs
from the vebus service are matched with a list of know hub-2
assistant IDs.

New versions of the hub-2 assistant will be recognized by the
vecan-mk2 service, which will publish a valid value at /Hub2
when such an assistant is found. This means that we do not
have to extend the list of assistants anymore when a new
assistant is released.