Implement auto-recharge. Also make state-machine a bit faster.
After reaching discharged state, if the SoC drops another 5%, we will start
charging from the grid. We stop charging when we reached the SoC at the time
we went into discharged state.

When the battery hits 0%, we always start charging from the grid, up to 3%.

Tests were also updated to test for the charger state.