Move BOL parameters into class properties. Ensure parameters are positive after applying the offset.
In the past, it was possible that a large offset value during debugging could
push a value negative, causing it to wrap on the Multi and become a large
positive value.

This change is also done in view of adding charge voltage compensation for
batteries that publish high static MaxChargeVoltage values and attempt
to control charging using current control only.