Add quirks feature, override BYD float voltage.
The main reason for this change is to avoid a sawtooth charging
pattern on BYD batteries. Instead of stopping charge completely,
we instead float it at 55V. This is implemented as a more generic
quirks feature so we can in theory add quirks for other batteries too.

Reorganise code so that current limits are determined and modified in a
single place.