Workaround for Sony/Murata very low CCL under discharge.
The Sony battery has a soft-start feature of 2A per module.
It sets this CCL the moment the battery drops out of a charging state,
for example while discharging. On systems with multiple phases, this throttles
all the Multis, even the ones that could charge. By ignoring the current limit
on these batteries, the user can set a limit on the CCGX dependent on his
battery size. This has been okayed by Sony: As long as we remain under 24A
per module it is safe.

The implementation simply sets a 1000A limit as long as the limit is above 0.
A zero limit is passed through unchanged.