change vebus system selection to lowest instance
Better because less random. And USB connected systems will automatically
be ignored if there is already one on the built-in port: USB connected
systems get a higher instance number.

Calculations:

- In case the Multi(s) are selected as the battery monitor, the system
already adds up DC volts and amps; which is fine: we assume they are
all on the same battery.

- Do not add up AC in or AC out values for the multiple systems: overview
should show the main system. Note that not adding it up will probably
mess up the total consumption power as shown on VRM dashboard.

https://github.com/victronenergy/venus/issues/339