Enable powering of the loads (or feeding in excess PV) inside a scheduled charging window.
Discharge is disabled inside a scheduled charge window, so that after the
target SOC is reached the battery will not start discharging until after
the scheduled charge period. Disabling discharge also causes hub4control
to disable feed-in, which means we cannot feed loads or feed in excess PV.
By simply changing the discharge limit to 1W instead of zero, we sidestep
this behaviour and feed-in is enabled while retaining the general non-discharge
behaviour.