Scheduled charging: Prioritise PV for loads
During scheduled charging discharge of the battery is disabled.  The intention
is to avoid premature discharge if the target SOC is reached before the
scheduled charge window ends.

In the past, any PV power simply continued to charge the battery, and upon
100% SOC the surplus went to the loads. This change reverses that: PV power
is used for the loads but limited in such a manner that the battery does not
discharge until the scheduled charge window ends.