Prefer VE.Bus over solarcharger voltage for SVS
In the rare case where a VE.Bus is available in the system but could not be
used for SOC (because other loads or chargers are present), systemcalc would
favour the solarcharger voltage over the vebus voltage, and sync the Multi with
the SolarCharger. We now always prefer the VE.Bus voltage and fall back to
other devices only if there is no VE.Bus.

https://github.com/victronenergy/venus/issues/448