batterydata: Create settings even for unused battery measurements.
A battery measurement can be unused, such as a starter battery on a BMV
or unused connections on a 3-way Skylla. In such cases we still want to
create a setting for the channel so that if a battery is connected later
the setting will work correctly in the GUI.

This also avoids a race condition where the battery voltage might not
be available at the time the device is detected. Observed with a Multi.