dvcc: Improve charge current limit
When disabling the DVCC charge current limit (user limit), the Multi
was not reset and would retain the old current limit until another was
set or a VE.Bus reset was performed.

This change writes a large value to /Dv/0/MaxChargeCurrent to reset the Multi.
Combined with changes in mk2-dbus this now resets the Multi to what was
configured with VEConfigure.

https://github.com/victronenergy/venus/issues/461
https://github.com/victronenergy/venus/issues/462