Change AcLost into more specific GridLost alarm
This simplifies it significantly. Timeout is shortened to 5 seconds (enough for
a vebus reset or startup). Code is now stateless. If the input type changes to
generator, we assume that indicates a loss of grid -- which is probably accurate.
We do this because there is no way to see if the non-active input (on a quattro)
has grid power, we can only react to conditions on the active input.