Add a 5% hysteresis on scheduled charging.
If the target SOC is reached before the end of the scheduled charging window,
the Multi will stop charging but delay discharge until the end of the window.
Within this window the SOC might drop by a small amount before the window
ends causing short pulses of charging every few minutes. Instead we now
add the well-known hysteresis of 5%, so that it will only reactivate
charging if the SOC drops by more than 5%.