Fix incorrect DC current when multi SOC disabled.
In fe2462a4 the assumption was made that if the extra current
from the solarchargers are written to /ExtraCurrent on the Multi,
them /Dc/0/Current (the VE.Bus current) will show the total with
the extra current already added. This assumption was incorrect.

The effect is that whenever the Battery Monitor is disabled in
VE.Configure, then systemcalc ignores the PV power and calculates
an incorrect Battery current.

This didn't have much of an effect until Venus 2.20 when /ExtraCurrent
disappeared from dbus if the batterymonitor was disabled.