SVS: Split control indicator paths for Multi and solar chargers.
In ESS systems the solar chargers are always synced with the Multi, but the
Multi is only synced if SVS is on. By using separate indicators it is clearer
what is going on and whether the battery is used.

The control indicators aren't presently used in the GUI. This is purely
for development and support.