Restore the slower speed of vebus dc-current tracking.
The VE.Bus DC current is tracked and added on to the charge current limit we
receive from an intelligent battery. This current is low-pass filtered to avoid
jostling solar chargers about. In 97ecfe the speed of this filter was increased
and also capped to avoid slower back-off. This change restores the speed
of the filter, but retains the capping behaviour.