allow quirks to raise the charge voltage
This adds support for the BYD Premium battery.

Previously there was a blanket clamp in place to ensure
that a quirk could not raise the charge voltage. The idea
was that we could never accidentally raise the charge voltage.
For the BYD Premium battery, it is necessary to raise the charge
voltage as it sets it 2V lower than the current battery voltage,
causing the Multi to discharge it.

The other quirks already clamp the voltage, so the blanket clamp
is not required.