deal with multi compact that shows disconnected when turned off
Some Multi models show that the input is disconnected (rather than invalid)
when turned off. This incorrectly causes the grid alarm to be raised. When
turning off the Multi remotely, there may also be a brief moment when
the Multi indicates the input to be disconnected just before it turns off.

Additionally increase the debouncing timeout to deal with false alarms
when the Multi is first turned on.