use extended state to avoid false grid alarms
On newer firmware, the Multi can report that it is in the process
of connecting to the grid. We can use this to avoid raising a grid
alarm where the grid is present but the Multi hasn't connected yet,
for example where a waiting period is mandated by the grid code.