broadcast vecan BMS voltage/temperature back to vecan
Normally we avoid writing a voltage or temperature reading back to the place
that it came from. In the case of vecan, where it is broadcast to all
devices on the bus, this was done by not forwarding anything from a vecan
device. It is however safe to forward it if the source is a battery. So
add an exception for that.