Don't send charge current limit to solar chargers with different N2kDeviceInstance than ourselves.
By default the device instance of the GX-device is zero, and presently it cannot
be changed. We therefore simply don't forward charge current limits to ve.can
chargers with a non-zero N2kDeviceInstance value.

This is the same as is done for voltage and network-mode in the
com.victronenergy.vecan service.

https://github.com/victronenergy/venus-private/issues/124