only check firmware version of solar chargers once
Solar chargers need a firmware version new enough to
work well with DVCC, so we raise an alarm if the firmware
version is too low. This should ideally be done using a different
method -- like capabilities -- but it is also unnecessary to do
this test every 3 seconds. The firmware cannot downgrade without
the solarcharger service leaving D-Bus and rejoining. Hence
we need only do this test once.

This also works around the problem that corruption occasionally causes
the firmware version to change for a second, causing intermittent false
alarms.

https://github.com/victronenergy/venus-private/issues/159