MFD: Avoid race using Multi as battery measurement
The general rule is that if /Dc/0/Voltage is invalid, then
a measurement cannot be used. This rule may affect the Multi
if the vebus service don't have /Dc/0/Voltage immediately after
becoming available on dbus. This may cause a delay in the multi
showing up as a battery measurement for MFD integration, or it
may not show up at all. But since the Multi cannot operate without
a battery anyway, this general rule is of no use and we can simply
always use the Multi as a battery measurement.

https://github.com/victronenergy/venus/issues/802