simplify bms selection
To make BMS selection the same as for hub4control, only use the
selected battery monitor as BMS (if it is a BMS).

This changes the earlier behaviour where the battery/BMS with the
lowest DeviceInstance was used regardless of the selected battery
monitor. This change breaks using a BMV as a battery manager and
still using the BMS for BOL parameters. A solution for this will
have to be found later, if at all.

https://github.com/victronenergy/venus/issues/783