Fix BmsControl indicator
The BmsControl control path is used so the Gui can indicate when the
BMS is in control. Until now it would only indicate BMS control if a
Multi was controlled. But in a system without a Multi, only solar
chargers or the new inverter RS, BMS control is applicable even without
a Multi. Therefore ammend this indication to also show when solar
chargers or an inverter is controlled.

https://github.com/victronenergy/venus/issues/803