shuts down solarchargers if CCL=0 for supported batteries
Because many batteries send CCL=0 when they are full, and not
necessarily because there is an emergency, we don't stop solar
chargers when we receive CCL=0. Instead we limit solar chargers
to what the loads require so that the batteries don't receive a
significant charge current.

This changeset causes no change in behaviour. It paves the way for
quirks to override this behaviour, and shut down charging on CCL=0.

This is required to support the upcoming Lynx Smart BMS.