add quirk for Lynx Smart BMS
For other BMSes, we don't stop the solarchargers completely if the BMS
sends CCL=0. That is because too many batteries send CCL=0 when they
are full, rather than when they are in an emergency situation. With
the Lynx Smart BMS CCL=0 is known to require a dead stop. So all
solar chargers are stopped.