fix crash when battery power is invalid
If for whatever reason /Dc/Battery/Power is invalid, the
systemstate delegate would cause a crash when it checks if
battery power is less than -30W.

This is a python3-related fix. In python2, None is always
less than anything except itself, so this would have gone
by without crashing.