improve missing bms detection
Instead of keeping a local flag in RAM, which will disappear
during a reboot/firmware update, instead use the BmsPresent
flag on the solar charger(s). If there is no BMS in the system
but the solar charger(s) expect one, then don't control the
solar charger(s). This will allow them to properly go into
error #67 when the BMS is lost in an ESS system.

More detail: In an ESS system, a charge voltage is still
available from the Multi, even when the BMS is missing.
If we copy that to the solar charger(s), we prevent them
from going into #67. This is an improvement on cae5a3.

Also: When there is a VE.Bus BMS v2, which also uses BmsPresent
on the solarchargers, skip this check. The Multi handles a missing
BMS correctly already by setting CVL=0.