dvcc: improve current limit distribution
Where solar chargers of different capacities were installed, then the
calculated current limit could exceed the capacity of the smaller solar
chargers. This would in turn causing an overall limit that is too low,
but only momentarily as it would be reset to the correct value 3
seconds later, causing oscilation.

No complaints in the 5 years it has been like this, so apparently not
a very common scenario.

Now we calculate a better balance that takes into account the size
of the solar charger.

https://github.com/victronenergy/venus/issues/890