fix PV not fed in during scheduled charge to 100%
Setting a discharge limit for 80% of the PV affects the setpoint of the Multi
calculated by hub4control. When hub4control uses mode SetpointIsMaxFeedIn,
as it almost always does when charging with DVCC on, this limits the amount of
overvoltage that can be fed in to 80% of what the solar chargers make, which
eventually causes the MPPTs to wind down.