optionally allow discharge during scheduled slot
Since it's introduction in Venus 2.30, Scheduled charging would
not allow discharging of the battery while inside a scheduled
slot, in order to avoid confusion about why the battery is being
discharged despite being in scheduled "charging". Only PV is allowed
to be used by loads.

In this commit there is now an option to allow the battery to
discharge if the SOC is above the limit. We will expose this in
the GUI as two options: Allowing self-consumption only from PV,
or from PV and battery.

https://github.com/victronenergy/venus/issues/970