allow discharge during scheduled charging
Since it's introduction in Venus 2.30, Scheduled charging would
not allow discharging of the battery while inside a scheduled
slot, in order to avoid confusion about why the battery is being
discharged despite being in scheduled "charging".

Now we allow discharge if a different scheduled slot arrives with
a lower target SOC. The problem of confusion will be resolved by
renaming the feature.