dvcc: Add support for 16-cell Pylontech batteries
The SOK SK48V100 battery uses a Pylontech BMS, but with 16 cells.
This change detects the higher charge voltage and allows the higher
charge voltage to be used instead. The 15-cell batteries, with a
charge voltage <54V are not affected.

A 16-cell Pylontech battery will also follow in the future.

https://github.com/victronenergy/venus/issues/1013