use AC-Input config to allocate Pv power summary
In the past there was a built-in assumption that PV-inverters
tied to AC-in-1 have their power allocated to PvOnGrid and AC-in-2
to PvOnGenset. This had (among others) the effect that if the Grid
was on AC-in-2, a common configuration, then the ESS overview would
not correctly show the PV-power, since it expects it on PvOnGrid.

Now we correctly interogate the AC input settings when allocating PV.

https://github.com/victronenergy/venus/issues/906