inverter/multi-rs: Fix DC system calculation
The handling of the sign of DC amps was incorrect for products with a
DC-measurement, causing the charging power of a Multi-RS or inverter-RS
to be added as a load. For older VE.Direct inverters, there is no DC
current measurement, and the load is estimated from the AC output power, which
has the opposite sign.

https://github.com/victronenergy/venus/issues/990