read state from active bms ito battery service
Where there is no VE.Bus system, we attempt to indicate
charge/discharge stopage by reading the paths on a managed
battery. Since introducing the feature to actively select
the BMS you want, the systemstate should get that info
from the selected BMS instead of the battery service,
and now it does.

https://github.com/victronenergy/venus/issues/901