vebus-bms: Don't control Multi charge current
When there is a VE.Bus BMS in the system, we set the charge current
of the solar chargers to a high value, or to zero, depending on
whether charge is allowed. Unfortunately we also copied the
remained of this value back to the Multi's charge current limit,
which is unnecessary since the Multi already listens to the VE.Bus
BMS.

Therefore, we skip the Multi charge current control when a VE.Bus
BMS is present.