ve-bus BMS v2: ctrl solarchargers when Multi off
The way the VE.Bus service is handled causes it to disappear
when /Connected = 0, which meant that we also stopped
controlling solar chargers when the Multi was switched off.

Now, we distinguish between the "Multi" (which is None of you
switch it off), and the VE.Bus service, which remains defined
as long as such a service exists.

Additionally, work around an issue on boards without an onboard
mk3. DVCC would start controlling the next VE.Bus service if you
switched off the first. Minor bug that will only affect mk-less
boards with two mk3-usb devices plugged in.