Don't show External Control when inverting
In Venus 2.93 we fixed an issue where external control was shown
on the VE.Bus state and this affected some customers. We instead
moved it to systemcalc, and inadvertently changed the behaviour
slightly. While technically correct (system is in external control
even when inverting), we want it to show "Inverting" as the system
state.

Therefore, revert previous behaviour where external control is
shown only when Multi is in Charge state (MainState = 9).

https://github.com/victronenergy/venus/issues/1073