dynamicess: charge/discharge speed depends on time
Adjust the speed of charging/discharging based on how much
time is allowed to get there. This is smoother than
importing/exporting the whole chunk of energy as fast as
possible and doing all the work in half the time allowed.

Additionally, the rate is recalculated when the SOC changes,
so if there was an interruption it will try to catch up.