dynamicess: fix oscillation around minsoc
When reaching the DESS minsoc, the system would go back into
self-consumption if the ESS minsoc was lower. This would
often result in slowly charging the battery again, which can
cause the battery to exceed DESS minsoc in a short while, causing
discharge to reactivate, and an oscillation around DESS minsoc.