activate recharge from slow charge
When the system is in slow charge, because of a battery that's been
in the discharged state for longer than 24 hours, but peak shaving
causes the SOC to drop even more, then activate recharge. When
minsoc is reached, go back to slow charge.

In the past, this would just get stuck in slow charge forever,
because it doesn't have enough charge power to reach the Discharged
state again.

https://github.com/victronenergy/venus/issues/1193