Always send charge current limit to Multi/Inv RS
Using the /State of Multi- and Inverter RS does not work the
same as a solar charger. The device is always on, even at night
when the sun is down, and in the case of the Multi it can also
charge from the grid and therefore needs the charge current
limit at all times anyway.

At the same time, some customers may configure the charge current
limit to be zero, and in such systems we can assume the charger is
off, and skip it when distributing charge current limits. This is
also required where multiple chargers are involved, as the splitting
algorithm doesn't deal kindly with a zero in the mix.

https://github.com/victronenergy/venus/issues/1186