dvcc: Remove dcdc converter from dvcc control
This includes SVS, STS and SCS, and acting as a temperature
service.

The decision was made that a DC/DC converter that is a charger
will be considered an alternator, and has the battery on the
output of the converter. This should be DVCC controlled. A DC/DC
converter that is a power supply (for some downstream equipment)
has the battery on the input and should not be controlled by DVCC.

https://github.com/victronenergy/venus-private/issues/310