reimplement internal charge current limit
The only way to limit charge current from the Multi when the Multi
has flag SetpointIsMaxFeedin set, is with the DC charge parameters.
This endpoint is here so that DynamicEss can limit the charge
current in order to reach the relevant SOC only at the end of
the time period.