dynamicess: keep charging if target is 100%
When a target of 100% is requested, don't stop charging at 100% and
go to idle mode. When feeding in of excess PV is not enabled, this
can lead to a condition where the battery is full, but the PV is not
used. By remaining in charge mode, the overvoltage feedin mechanism
will power loads, and the battery also has an opportunity to balance.