dvcc: Avoid deadlock charge current stuck low
If a managed battery disappears for a moment, and a charge current
limit was set manually, that limit can get stuck on the Multi,
because we write the limit to different places depending on whether
a bms is present or not. To fix this, always ensure that the non-bol
value is not less than the bol value.

https://github.com/victronenergy/venus/issues/1224