Sync with localsettings, add Automax setting
Sync min/max with localsettings, use /Settings/Gui/Gauges/Automax to switch between automatic
updating of min/max or setting manual limits.

support for different AC in scenarios:
1: AC in 1 is connected
2: AC in 2 is connected
3: No AC is connected

The maximum reported by Ac/Consumption/Current/Max depends on the active scenario
Because the system capability may be less when the AC input gets disconnected or
switches from grid to generator
So maintain a maximum for all three situations, and relay the correct one
to /Ac/Consumption/Current/Max for the gui to use.