hubtype: use multi delegate to find vebus instance
This is a simple sideways improvement, use the newer centralised
method of knowing which Multi to read hub information from. Since
an earlier version we no longer support DVCC on Multis not
connected to the built-in port, and this extends it to also not
getting the hub-type from a system we don't control.