hubtype: Don't default to Hub-1
If there is no Multi, many systems will show Hub-1 as long as
there is PV power present. That is not correct. Now we show
invalid instead.

In future we'll properly deal with the Multi-RS's precise hub
type, or some other method of showing ESS capacility rather than
a text string.