allow using a callable for gettext
Extend the mechanism for defining the output paths, where a
gettext option provides a way to format a value, so that it
can also take a callable, for more powerful mapping to a
good string.

Also make the gettext option, optional. Default is then to simply
return a string representation.