loadshedding: refactor reconnect margin
Reconnect margin is now the time AFTER the slot starts before
a reconnection is allowed. If the power does not fail, or
returns sooner than the end of the schedule, reconnection will
be attempted after this margin has passed.

This would typically be set to 30 minutes or more. If the power
has not failed within 30 minutes after the start of the schedule,
the schedule is probably wrong, and we can reconnect.

Similarly, if at least half an hour has passed, and the power
returns early, we can assume the schedule was cancelled, and
reconnect early.