improve dc system estimate, add system current
Add a path for DC system current. Also improve the estimate
by removing the alternator power from the DC system calculation.

This was initially added in 3290f4 by mbath, and removed in
fc9f4a because some people preferred to see alternator power
in their DC loads box, but now that we are moving to GUIv2
we can fix this properly.

https://github.com/victronenergy/gui-v2/issues/739