improve systemstate display under external control
In ac8a23 a change was made so that systems that are externally
controlled does not show random things as system state.
Unfortunately this left an edge case where a system that formerly
had ESS installed and was externally controlled would show the
incorrect state if ESS is later removed, with no way to correct
this other than reinstalling ESS and resetting the ESS mode.

This corrects that oversight and short-circuits the system state
display only if ESS is installed