Pylontech: Apply quirk also to 15-cell batteries
We have a quirk that prevents the charge current from being
lowered all the way to zero, and instead we use a lower
charge voltage. This quirk was applied only to the 24V
batteries, but since there are plenty of 15-cell batteries
that also send CCL=0, it is better to also apply it to those.