acinput: better support sans invertercharger
In systems without an Inverter/Charger, there is no information
about the inputs. We can however make educated guesses based on
the presence of a grid meter or a genset service. In that case,
populate the Ac/In structures with the grid/genset meter info
and assume the first one found is actually avtive. That will allow
better data to be shown on GUI-v2/VRM.

https://github.com/victronenergy/venus/issues/1311