restore 48V pylontech quirk
In df8bd3 the 48V battery quirk was changed to also avoid CCL=0
issues. Now that we know this may well cause more problems in the
field, it is perhaps wise not to change this now, and to restore
it to its previous version: CCL=0 is again passed to the Multi
as it is in Venus 3.40.

https://github.com/victronenergy/venus-private/issues/424