pylontech: experimental charge voltage algo
Experimental charge voltage algorithm that lowers the CVL when the
CCL is less than C/5. This is a high condition, and in this condition
the CVL is set to 14 times the lowest cell, plus the highest cell,
thereby drawing the voltage down to the lower value. If the BMS
asks for more than C/5, we flip this upside down: We ask for 14
times the highest cell, plus the lowest cell, thereby pulling all
the low cells up to the highest cell.

In short, this implements dynamic voltage control for Pylontech
US-series batteries, as long as they run relatively new firmware.
With older firmware, the charge voltage is simply lowered to 3.45V
per cell when the BMS lowers CCL.

https://github.com/victronenergy/venus-private/issues/424