dvcc: handle inverterchargers separately
Until now, inverter/chargers (Multi-RS) was treated simply as a
solar charger. This is a refactoring to handle them separately,
so we can treat them differently later. This is necessary in case
we want to control solar chargers with an external voltage that
comes from an inverter/charger (avoid loops), and also later when
it becomes possible to control the AC-charging separately.

No functional change in this commit.