dvcc: Send RS solar offset to other chargers
In a system where there are VE.Direct solar chargers, add the
solar offset from the RS to the charge voltage to create an
offset for feeding PV into the grid or to loads.

Also, in a system without a managed battery, use the charge
voltage from the RS, plus the solar offset.

https://github.com/victronenergy/venus-private/issues/368