pylontech: further voltage control improvement
This only works if the battery firmware is new enough and it
communicates the highest cell coltage. Otherwise the old
behaviour of using 52.4V is retained.

The charge voltage is adjusted such that the highest cell is held
between 3.485V and 3.5V. It automatically compensates for the
overvoltage offset that is added for feeding in excess DC-coupled
PV.

This is sufficient for the SOC to reset to 100%, while it should
also prevent high voltage warnings in batteries that aren't
full balanced.

https://github.com/victronenergy/venus-private/issues/424