add config option for system with no AC-in loads
Some systems have a grid meter because they want vector registration on
energy counting, but there aren't any actual loads in front of the
inverter/charger. In such systems, showing the AC input loads makes
little sense and can in fact be noisy, since a small amount of power
is almost always reflected there.

This setting skips the publishing of ConsumptionOnInput, causing
the various GUIs to show that correctly.

https://github.com/victronenergy/venus/issues/1328