batteryservice: deal with garbage from batteries
For reasons that could not be quite worked out, a battery
monitor managed to get a broken value into systemcalc,
as product name, custom name, or deviceinstance, and this
caused an exception in the list of available BMSes.

Therefore, filter the list carefully and only publish items
that have a string for a name, and an integer for a
deviceinstance.

https://github.com/victronenergy/venus-private/issues/397