DESS: check ESS assistant before SOC
For DESS to be available, we need an SOC as well as the ESS
assistant installed on the Multi. The SOC is sometimes invalid
at the same time that the ESS assistant is not present, such
as when the Multi is off. So checking for that first makes
errorcode 4 show up instead of errorcode 1, which is the
main reason. So we swap them to be more accurate.

https://github.com/victronenergy/venus/issues/1368