dvcc, RS: fix charge voltage somtimes 0 or 0.4
In an RS system without a managed battery, when the charge voltage
is not (yet) defined, but the solar offset is, the system might
add them together and use 0V or 0.4V as the charge voltage. This
will shut down all non-VE.Can chargers for a short while.

Avoid this by explicitly checking for a charge voltage before adding
them together.