fix for BYD batteries with large imbalance
Some BYD batteries (Flex, Premium) have a dynamic charge voltage.
Which means they rarely need to send CCL=0. But when they do, the
old quirk for the B-Box would pin the charge voltage to 55V, which
had the unfortunate side effect that in a poorly balanced battery
it can actually raise the charge voltage.

So now we set it to 55V, or what the BMS asks for, whichever is
lower.