dess: disconnected vebus system would break dess
On any system that once had a VE.Bus inverter/charger connected,
this will be remembered and a disconnected vebus service will show
up. This causes DESS to try to use the vebus system, but since it
does not have the ESS assistant it then indicates that DESS is not
supported.

Now, we check that the device has /Connected=1, so it will skip the
disconnected ve.bus system.