allow SOC to sync across VE.Can ports
If a VE.Can BMS is on a different port than Multi RS devices,
then no SOC syncing was done. The system would refuse to sync
a VE.Can derived SOC back to VE.Can, to avoid a loop.

If the source of the SOC is a battery, however, it is safe to
sync it back, and this also allows syncing across different
CAN ports.

Therefore, sync SOC back to VE.Can if the source is a battery.