dvcc: fix inconsistencies in charger control
In the older code, if CCL=0 and it was the kind of battery where
that means an absolute stop, it would stop ALL chargers, instead
of just the ones that are actually active, and had the correct
N2kDeviceInstance. This wasn't necessarily wrong, but it was
inconsistent: Sometimes it controls all chargers, sometimes just
a selected group.

Now it will control only active chargers (state is not off), with
the correct N2kDeviceInstance, so that chargers in different groups
are not stopped on CCL=0.