dvcc: Also control charge current of other multis
When there is more than one Multi in the system, only the charge
current of the first was controlled by DVCC. Now we will also
send the same charge current as calculated for the first Multi
to the second Multi. No balancing is done.

This is for systems where one Multi might be connected to a
genset while another is connected to the grid. Not completely
uncommon. This allows the Multi connected to the genset to
charge according to the DVCC limit.

This is NOT meant for use with both Multis connected to the grid.
In that case, the battery may be charged with more current than
requested.

https://github.com/victronenergy/venus-private/issues/564