show ess self consumption state on dbus
This is also known as the BatteryLife state, but more accurately
it shows the ESS state in systems with a Multiplus. If there is
no Multi with ESS assistant in the system, then this will instead
now show invalid, so that it will not be logged to VRM.

Additionally, give /Control/ActiveSocLimit the same treatment so it
is invalid if not used.

https://github.com/victronenergy/venus-private/issues/572